Validation Checks
I was using this model for my project, but I faced a few issues with it, so I changed the code and solved my problem.
First, in "classification.py" file we need to check if the "temp_rows" list is not empty, otherwise it causes error.
Second, in "classification.sh" file by doing that change, I could use this model in google colab.